requirements.txt
streamlit
pdfplumber
pdf2image
pytesseract
opencv-python-headless
numpy
Pillow
google-genai